+++ clang-tools-extra/clang-tidy/performance/FasterStringOperationCheck.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "FasterStringFindCheck.h"
+#include "FasterStringOperationCheck.h"
 #include "../utils/OptionsUtils.h"
 #include "clang/AST/ASTContext.h"
 #include "clang/ASTMatchers/ASTMatchFinder.h"
@@ -42,19 +42,20 @@ makeCharacterLiteral(const StringLiteral *Literal) {
   return Result;
 }
 
-FasterStringFindCheck::FasterStringFindCheck(StringRef Name,
-                                             ClangTidyContext *Context)
+FasterStringOperationCheck::FasterStringOperationCheck(
+    StringRef Name, ClangTidyContext *Context)
     : ClangTidyCheck(Name, Context),
       StringLikeClasses(utils::options::parseStringList(
           Options.get("StringLikeClasses",
                       "::std::basic_string;::std::basic_string_view"))) {}
 
-void FasterStringFindCheck::storeOptions(ClangTidyOptions::OptionMap &Opts) {
+void FasterStringOperationCheck::storeOptions(
+    ClangTidyOptions::OptionMap &Opts) {
   Options.store(Opts, "StringLikeClasses",
                 utils::options::serializeStringList(StringLikeClasses));
 }
 
-void FasterStringFindCheck::registerMatchers(MatchFinder *Finder) {
+void FasterStringOperationCheck::registerMatchers(MatchFinder *Finder) {
   const auto SingleChar =
       ignoringParenCasts(stringLiteral(hasSize(1)).bind("literal"));
 
@@ -78,7 +79,7 @@ void FasterStringFindCheck::registerMatchers(MatchFinder *Finder) {
                      this);
 }
 
-void FasterStringFindCheck::check(const MatchFinder::MatchResult &Result) {
+void FasterStringOperationCheck::check(const MatchFinder::MatchResult &Result) {
   const auto *Literal = Result.Nodes.getNodeAs<StringLiteral>("literal");
   const auto *FindFunc = Result.Nodes.getNodeAs<FunctionDecl>("func");